pyproject.toml
[project]
name = "accentra"
version = "0.1.0"
description = "User management"
authors = [
    { name = "Thomas Wolter", email = "support@riskary.de" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.118.1",
    "opentelemetry-exporter-otlp>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.27",
    "dramatiq>=1.16.0",
    "redis>=6.4.0",
    "tenauth @ git+https://github.com/thwolter/fde-tenauth.git@main",
    "pydantic[email]>=2.12.0",
    "uvicorn>=0.37.0",
    "pyjwt>=2.10.1",
]
[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "commitizen>=4.9.1",
    "dramatiq[watch]>=1.18.0",
    "isort>=6.1.0",
    "mkdocs>=1.6.1",
    "pre-commit>=4.3.0",
    "pyrefly>=0.36.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.0",
    "ssort>=0.15.0",
    "typer>=0.19.2",
    "vulture>=2.14",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = [
    "**/node_modules",
    "**/__pycache__",
    "**/*venv/**/*",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
